==> 6_BIOVARS_NCCS_DATA_MODELS.R <==
################################################################################
################################################################################
###The script produces 19 biovars using dismo package. The function takes monthly P,
###Tasmax and tasmin as input to get annual biovars values as an output. Here I used monthly
###data and then I aggregated the annual result to get individual model values for the 30-year periods.
################################################################################
################################################################################

library(dismo)
library(raster)
library(sf)
library(terra)
library(dplyr)

memory.limit(size=560000)
rasterOptions(progress = "text")
rasterOptions(timer = TRUE)
rasterOptions(chunksize = 1e+10)
rasterOptions(maxmemory = Inf)
rasterOptions(memfrac = 0.9)
rasterOptions(tmptime = 72)

rm(list=ls())
gc()


# Preparing input data and setting variables for looping ----------------

input_dir <- "2_MONTHLY_STACKS"
output_dir <- "3_ANNUAL_STACKS"
output_dir_dataframe <- "3A_ANNUAL_STACKS_DF"
output_dir_summary <- "4_ANNUAL_STACKS_SUMMARY"
output_dir_dataframe_agg <- "4A_ANNUAL_STACKS_SUMMARY_DF"

grassland <- st_read("GIS/WWF_TERR_ECO/wwf_tropical_montane.shp")
grassland_dis <- grassland %>% group_by(ECO_NAME, ECO_ID, G200_REGIO) %>% summarize()

grassland_v <- vect("GIS/WWF_TERR_ECO/wwf_tropical_montane.shp")

experiments <- c("historical", "ssp585")
models <- dir(input_dir)

# Looping by model/experiment/variable -------------------------------------------------

mod <- 1
for (mod in 1:length(models)) {
  
  print(paste0("-----","(", Sys.time(), ")", " Model: ", models[mod], " - Start processing biovars!"))
  
  exp <- 1
  for (exp in 1:length(experiments)) {
    
    print(paste0("-----","(", Sys.time(), ")", " Experiment: ", experiments[exp], " - Start processing biovars!"))
    
    # Load pr, tasmin and tasmax
    # Select just the bands for the 30-years period and crop to the study area

    pr <- rast(list.files(file.path(input_dir, models[mod], "pr", experiments[exp]), pattern = "\\.RDS$|\\.grd$|\\.tif$", full.names = TRUE))
    pr <- pr[[(nlyr(pr) - ((30*12)-1)):nlyr(pr)]]
    pr <- terra::mask(pr, grassland_v, touches = TRUE)
    pr <- stack(pr)
    
    tasmin <- rast(list.files(file.path(input_dir, models[mod], "tasmin", experiments[exp]), pattern = "\\.RDS$|\\.grd$|\\.tif$", full.names = TRUE))
    tasmin <- tasmin[[(nlyr(tasmin) - ((30*12)-1)):nlyr(tasmin)]]
    tasmin <- terra::mask(tasmin, grassland_v, touches = TRUE)
    tasmin <- stack(tasmin)
    
    tasmax <- rast(list.files(file.path(input_dir, models[mod], "tasmax", experiments[exp]), pattern = "\\.RDS$|\\.grd$|\\.tif$", full.names = TRUE))
    tasmax <- tasmax[[(nlyr(tasmax) - ((30*12)-1)):nlyr(tasmax)]]
    tasmax <- terra::mask(tasmax, grassland_v, touches = TRUE)
    tasmax <- stack(tasmax)

    # Set control variables to iterate by year
    years <- 1:(nlayers(pr)/12)
    start <- 1
    end <- 12
    
    if(experiments[exp] == "historical"){
      years_vector <- 1985:2014
    } else {
      years_vector <- 2071:2100
    }
    
    bvars <- stack() #empty stack to store results from the loop
    
    for (year in 1:length(years)) {
      
      print(paste0("-----","(", Sys.time(), ")", " Processing biovars for year : ", years_vector[year]))
      print(paste0("Start index: ", start))
      print(paste0("End index: ", end))
      
      timeStart<- proc.time()
      
      prec <- pr[[start:end]]

      tmin <- tasmin[[start:end]]

      tmax <- tasmax[[start:end]]

      
      # Stack the annual biovars generated by looping
      bvars <- stack(bvars, biovars(prec = prec,
                                    tmin = tmin,
                                    tmax = tmax,
                                    na.rm = TRUE))
      
      # Increase values of control variables
      start <- start + 12
      end <- end + 12
      
      proc.time() - timeStart
      
    }
    
    rm(pr, tasmax, tasmin)
    gc()
    
    #Save results in disk using terra (faster)
    bvars_rast <- terra::rast(bvars)
    
    if (!dir.exists(file.path(output_dir, models[mod], experiments[exp]))) {
      dir.create(file.path(output_dir, models[mod], experiments[exp]), recursive = TRUE)
    }
    
    terra::writeRaster(x = bvars_rast,
                       filename = file.path(output_dir, models[mod], experiments[exp], paste0("biovars_", experiments[exp], ".tif")),
                       filetype = "GTiff",
                       datatype = "FLT4S",
                       overwrite = TRUE,
                       memfrac = 0.9)

    #Reload the biovars to work from disk, not from memory
    bvars_rast <- terra::rast(file.path(output_dir, models[mod], experiments[exp], paste0("biovars_", experiments[exp], ".tif")))

    # Looping by biovar to produce separate rasters by biovar/experiment (annual basis)
    bvar <- 1
    for (bvar in 1:19) {
      
      print(paste0("---","(", Sys.time(), ")", " Processing biovar: ", bvar))
      
      if (!dir.exists(file.path(output_dir, models[mod], paste0("biovar", bvar), experiments[exp]))) {
        dir.create(file.path(output_dir, models[mod], paste0("biovar", bvar), experiments[exp]), recursive = TRUE)
      }
      
      print(paste0("-","(", Sys.time(), ")", " Exporting biovar: ", bvar))
      
      # Create a subset of the biovars to produce separate rasters by biovar
      bvar_subset <- terra::subset(bvars_rast,
                                   subset = seq(from = bvar, by = 19, length.out = length(years)))
      
      names(bvar_subset) <- paste0("biovar", bvar, ".", years_vector)
      
      # Save results in disk using terra (faster)
      terra::writeRaster(x = bvar_subset,
                         filename = file.path(output_dir, models[mod], paste0("biovar", bvar), experiments[exp], paste0("biovar", bvar, ".tif")),
                         filetype = "GTiff",
                         datatype = "FLT4S",
                         overwrite = TRUE,
                         memfrac = 0.9)
      
      print(paste0("-","(", Sys.time(), ")", " Exporting dataframe for biovar: ", bvar))
      
      # Reload the biovar to work from disk, not from memory
      bvar_subset <- raster::stack(file.path(output_dir, models[mod], experiments[exp], paste0("biovar", bvar, ".tif")))
      
      # Extract each biovar values using TAE polygons to a DF
      data.biovar <- raster::extract(bvar_subset, grassland_dis,
                                     weights = TRUE,
                                     df = TRUE,
                                     exact = TRUE,
                                     small = TRUE,
                                     normalizeWeights = TRUE,
                                     cellnumbers = TRUE)
      
      if (!dir.exists(file.path(output_dir_dataframe, models[mod], paste0("biovar", bvar), experiments[exp]))) {
        dir.create(file.path(output_dir_dataframe, models[mod], paste0("biovar", bvar), experiments[exp]), recursive = TRUE)
      }
      
      # Save DFs
      write.csv(data.biovar, file.path(output_dir_dataframe, models[mod], paste0("biovar", bvar), experiments[exp], paste0("biovar", bvar, ".csv")))
      
      print(paste0("-","(", Sys.time(), ")", " dataframe for biovar: ", bvar, " exported!"))
      
      print(paste0("-","(", Sys.time(), ")", " Exporting aggregated biovar: ", bvar))
      
      if (experiments[exp] == "historical") {
        

        y <- "_1985-2014"
        
      } else {
        
        y <- "_2071-2100"
        
      }
      
      if (!dir.exists(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), experiments[exp]))) {
        dir.create(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), experiments[exp]), recursive = TRUE)
      }
      
      # Aggregate rasters to produce the 30-year period rasters
      r.biovar.agg <- raster::stackApply(bvar_subset,
                                         indices = rep(1, nlayers(bvar_subset)),
                                         fun = mean,
                                         na.rm = TRUE)
      
      #Save 30-year period aggregated rasters
      raster::writeRaster(r.biovar.agg,
                          filename = file.path(output_dir_summary, models[mod], paste0("biovar", bvar), experiments[exp], paste0("biovar", bvar, "_mod", y, ".tif")),
                          datatype = "FLT4S",
                          format = "GTiff",
                          overwrite = TRUE)
      
      print(paste0("-","(", Sys.time(), ")", " Aggregated biovar: ", bvar, " exported!"))
      
      print(paste0("-","(", Sys.time(), ")", " Exporting dataframe for aggregated biovar: ", bvar))
      
      # Extract DF for the 30-year aggregated rasters
      data.biovar.agg <- raster::extract(r.biovar.agg, grassland_dis,
                                         weights = TRUE,
                                         df = TRUE,
                                         exact = TRUE,
                                         small = TRUE,
                                         normalizeWeights = TRUE,
                                         cellnumbers = TRUE)
      
      if (!dir.exists(file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), experiments[exp]))) {
        dir.create(file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), experiments[exp]), recursive = TRUE)
      }
      
      #Save aggregated extracted df
      write.csv(data.biovar.agg, file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), experiments[exp], paste0("biovar", bvar, "_mod", ".csv")))
      
      print(paste0("-","(", Sys.time(), ")", " Dataframe for aggregated biovar: ", bvar, " exported!"))
      
    }
  }
}



# Create difference among the two 30-year period rasters for each biovar and then extract DFs --------

for (mod in 1:length(models)) {

  for (biovar in 1:19) {

    r.norm.h <- stack(list.files(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), "historical"),
                                 full.names = TRUE, recursive = TRUE, pattern = "\\.tif$"))

    r.norm.f <- stack(list.files(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), "ssp585"),
                                 full.names = TRUE, recursive = TRUE, pattern = "\\.tif$"))

    r.norm.dif <- r.norm.f - r.norm.h
    
    if (!dir.exists(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), "diff"))) {
      dir.create(file.path(output_dir_summary, models[mod], paste0("biovar", bvar), "diff"), recursive = TRUE)
    }

    raster::writeRaster(r.norm.dif,
                        filename = file.path(output_dir_summary, models[mod], paste0("biovar", bvar), "diff", paste0("biovar", biovar, "_mod_diff.tif")),
                        format = "GTiff",
                        datatype = "FLT4S",
                        overwrite = TRUE)

    data.dif <- raster::extract(r.norm.dif, grassland_dis,
                                weights = TRUE,
                                df = TRUE,
                                exact = TRUE,
                                small = TRUE,
                                normalizeWeights = TRUE,
                                cellnumbers = TRUE)
    
    if (!dir.exists(file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), "diff"))) {
      dir.create(file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), "diff"), recursive = TRUE)
    }

    write.csv(data.dif, file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", bvar), "diff", paste0("biovar", biovar, "_mod_diff.csv")))

    rm(r.norm.h, r.norm.f, r.norm.dif, data.dif)
    gc()

  }

}


# Summarize data for model/variable/ecoregion -----------------------------

models <- dir(input_dir)

experiments <- c("historical", "ssp585", "diff")

biovars <- 1:19

std.error <- function(x) sd(x)/sqrt(length(x))

for (exp in 1:length(experiments)) {
  
  for (biovar in biovars) {
    
    if (!dir.exists(file.path(output_dir_dataframe_agg, "ENSEMBLE", paste0("biovar", biovar), experiments[exp]))) {
      dir.create(file.path(output_dir_dataframe_agg, "ENSEMBLE", paste0("biovar", biovar), experiments[exp]), recursive = TRUE)}
    
    df.list <- list()
    
    for (mod in 1:length(models)) {
      
      df.mod <- read.csv(list.files(file.path(output_dir_dataframe_agg, models[mod], paste0("biovar", biovar), experiments[exp]), full.names = TRUE))
      names(df.mod) <- c("X", "ID", "cell", "X1", "weight")
      df.mod <- df.mod %>%
        group_by(ID) %>%
        summarise(wmean = weighted.mean(x = X1, w = weight))
      names(df.mod) <- c("ID", models[mod])
      
      df.list[[mod]] <- df.mod
    }
    
    df.ensemble <- do.call(cbind, df.list)
    df.ensemble <- df.ensemble[, !duplicated(colnames(df.ensemble))]
    df.ensemble$mean <- rowMeans(df.ensemble[2:7])
    df.ensemble$standard_dev <- apply(df.ensemble[2:7], 1, sd)
    df.ensemble$standard_err <- apply(df.ensemble[2:7], 1, std.error)
    
    write.csv(df.ensemble, file.path(output_dir_dataframe_agg, "ENSEMBLE", paste0("biovar", biovar), experiments[exp], paste0("biovar", biovar, "_ens_summary_", experiments[exp], ".csv")))
    
  }
  
}
